requirements.txt
asgiref==3.8.1
Django==4.2.10
sqlparse==0.5.2
tzdata==2024.2
Pillow==10.4.0
psycopg2-binary
dj-database-url